refactor: 安装链归 src/lib——install/errclass/pnpm 是宿主自动链实现层，非工具库
dsh_install 已非工具形态：D6 零干预后依赖安装由宿主自动链（onload 自动装/状态机）主导，
tools/dsh-install.js 只是 Agent 手动入口薄壳（调 g.installDeps/verifyDeps）。install/
errclass/pnpm 主体服务宿主 → src/tools/lib 归 src/lib（与 bus/lifecycle/profile-seed 同层）；
引用修正（index/lifecycle/dsh-run/pnpm 内部/tests）。src/lib = 宿主核心；src/tools/lib
剩 state/config（宿主工具共用）与 protocol/wake（会话工具链）。27/27 + build 通过。
